niffler-e-2-e-tests-python/requirements.txt
pytest
pytest-playwright
email-validator
python-dotenv
allure-pytest
requests
pydantic
faker